+++ libpod/network/files.go
@@ -11,6 +11,7 @@ import (
 	"github.com/containernetworking/plugins/plugins/ipam/host-local/backend/allocator"
 	"github.com/containers/common/pkg/config"
 	"github.com/containers/podman/v3/libpod/define"
+	"github.com/containers/podman/v3/pkg/network"
 	"github.com/pkg/errors"
 	"github.com/sirupsen/logrus"
 )
@@ -67,7 +68,7 @@ func GetCNIConfigPathByNameOrID(config *config.Config, name string) (string, err
 		if conf.Name == name {
 			return confFile, nil
 		}
-		if strings.HasPrefix(GetNetworkID(conf.Name), name) {
+		if strings.HasPrefix(network.GetNetworkID(conf.Name), name) {
 			idMatch++
 			file = confFile
 		}
@@ -90,16 +91,6 @@ func ReadRawCNIConfByNameOrID(config *config.Config, name string) ([]byte, error
 	}
 	b, err := ioutil.ReadFile(confFile)
 	return b, err
-}
-
-// GetCNIPlugins returns a list of plugins that a given network
-// has in the form of a string
-func GetCNIPlugins(list *libcni.NetworkConfigList) string {
-	plugins := make([]string, 0, len(list.Plugins))
-	for _, plug := range list.Plugins {
-		plugins = append(plugins, plug.Network.Type)
-	}
-	return strings.Join(plugins, ",")
 }
 
 // GetNetworkLabels returns a list of labels as a string
