+++ libpod/network/network.go
@@ -1,8 +1,6 @@
 package network
 
 import (
-	"crypto/sha256"
-	"encoding/hex"
 	"encoding/json"
 	"net"
 	"os"
@@ -243,13 +241,6 @@ func Exists(config *config.Config, name string) (bool, error) {
 		return false, err
 	}
 	return true, nil
-}
-
-// GetNetworkID return the network ID for a given name.
-// It is just the sha256 hash but this should be good enough.
-func GetNetworkID(name string) string {
-	hash := sha256.Sum256([]byte(name))
-	return hex.EncodeToString(hash[:])
 }
 
 // PruneNetworks removes networks that are not being used and that is not the default
